+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   Unifies usage with pure dropwizard (no additional `@Provider` annotation required). (#265)
     - New option InstallerOptions.JerseyExtensionsRecognizedByType could disable new behaviour
 * Support ModelProcessor jersey extension installation (#186)
+* Add extensions help: .printExtensionsHelp() showing extension signs recognized by installers (in recognition order)
+  - Custom installers could participate in report by overriding FeatureInstaller.getRecognizableSigns()
+    (default interface method).
 
 ### 5.6.1 (2022-07-02)
 * Update dropwizard to 2.1.1 (fixes java 8 issue by allowing afterburner usage)